New package structure and class names
