--- delphesprocessor/__init__.py
@@ -1,1 +1,0 @@
-from madminer.__version__ import __version__
