+++ environment.yml
@@ -3,7 +3,7 @@ channels:
 - pytorch
 - anaconda
 - defaults
-- conda-forge
+- conda-ml
 dependencies:
 - python=3.6.*
 - numpy
