+++ madminer/__version__.py
@@ -1,1 +1,1 @@
-__version__ = '2018.08.31'
+__version__ = '2018.09.17'
