+++ madminer/utils/analysis.py
