+++ madminer/utils/ml/models/batch_norm.py
