+++ madminer/utils/ml/score_trainer.py
