+++ madminer/utils/plots.py
